LPD-42620 Move the logic to private methods and delete the FriendlyUrlHistoryDTOConverter
